src/loader/requirements.txt
greenlet==3.0.3
psycopg2-binary==2.9.9
PyYAML==6.0.2
SQLAlchemy==2.0.32
typing_extensions==4.12.2
